GafferScene : Move InteractiveRender out of Preview namespace
This entirely replaces the original InteractiveRender node, which was unused following the demise of GafferRenderMan.
