+++ src/GafferSceneModule/RenderBinding.cpp
@@ -47,7 +47,6 @@
 #include "GafferScene/SceneProcedural.h"
 #include "GafferScene/ExecutableRender.h"
 #include "GafferScene/Preview/Render.h"
-#include "GafferScene/Preview/InteractiveRender.h"
 #include "GafferScene/Private/IECoreScenePreview/Renderer.h"
 #include "GafferScene/Private/IECoreScenePreview/Procedural.h"
 
@@ -106,11 +105,6 @@ class ExecutableRenderWrapper : public TaskNodeWrapper<ExecutableRender>
 };
 
 ContextPtr interactiveRenderGetContext( InteractiveRender &r )
-{
-	return r.getContext();
-}
-
-ContextPtr previewInteractiveRenderGetContext( Preview::InteractiveRender &r )
 {
 	return r.getContext();
 }
@@ -230,7 +224,8 @@ void GafferSceneModule::bindRender()
 	{
 		scope s = GafferBindings::NodeClass<InteractiveRender>()
 			.def( "getContext", &interactiveRenderGetContext )
-			.def( "setContext", &InteractiveRender::setContext );
+			.def( "setContext", &InteractiveRender::setContext )
+		;
 
 		enum_<InteractiveRender::State>( "State" )
 			.value( "Stopped", InteractiveRender::Stopped )
@@ -244,19 +239,6 @@ void GafferSceneModule::bindRender()
 		scope().attr( "Preview" ) = previewModule;
 
 		scope previewScope( previewModule );
-
-		{
-			scope s = GafferBindings::NodeClass<GafferScene::Preview::InteractiveRender>()
-				.def( "getContext", &previewInteractiveRenderGetContext )
-				.def( "setContext", &GafferScene::Preview::InteractiveRender::setContext )
-			;
-
-			enum_<GafferScene::Preview::InteractiveRender::State>( "State" )
-				.value( "Stopped", GafferScene::Preview::InteractiveRender::Stopped )
-				.value( "Running", GafferScene::Preview::InteractiveRender::Running )
-				.value( "Paused", GafferScene::Preview::InteractiveRender::Paused )
-			;
-		}
 
 		{
 			scope s = TaskNodeClass<GafferScene::Preview::Render>();
